+++ tests/image/test_image.py
@@ -8,14 +8,10 @@ import pathlib
 from PIL import Image
 from resizeimage.imageexceptions import ImageSizeError
 
-from zimscraperlib.imaging import (
-    get_colors,
-    is_hex_color,
-    resize_image,
-    create_favicon,
-    convert_image,
-    save_image,
-)
+from zimscraperlib.image.probing import get_colors, is_hex_color
+from zimscraperlib.image.transformation import resize_image
+from zimscraperlib.image.convertion import create_favicon, convert_image
+from zimscraperlib.image import save_image
 
 
 def get_image_size(fpath):
@@ -86,7 +82,7 @@ def test_resize_thumbnail(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 100, 50
-    resize_image(src, width, height, to=dst, method="thumbnail")
+    resize_image(src, width, height, dst=dst, method="thumbnail")
     tw, th = get_image_size(dst)
     assert tw <= width
     assert th <= height
@@ -99,7 +95,7 @@ def test_resize_width(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 100, 50
-    resize_image(src, width, height, to=dst, method="width")
+    resize_image(src, width, height, dst=dst, method="width")
     tw, _ = get_image_size(dst)
     assert tw == width
 
@@ -111,7 +107,7 @@ def test_resize_height(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 100, 50
-    resize_image(src, width, height, to=dst, method="height")
+    resize_image(src, width, height, dst=dst, method="height")
     _, th = get_image_size(dst)
     assert th == height
 
@@ -123,7 +119,7 @@ def test_resize_crop(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 5, 50
-    resize_image(src, width, height, to=dst, method="crop")
+    resize_image(src, width, height, dst=dst, method="crop")
     tw, th = get_image_size(dst)
     assert tw == width
     assert th == height
@@ -136,7 +132,7 @@ def test_resize_cover(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 5, 50
-    resize_image(src, width, height, to=dst, method="cover")
+    resize_image(src, width, height, dst=dst, method="cover")
     tw, th = get_image_size(dst)
     assert tw == width
     assert th == height
@@ -149,7 +145,7 @@ def test_resize_contain(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 5, 50
-    resize_image(src, width, height, to=dst, method="contain")
+    resize_image(src, width, height, dst=dst, method="contain")
     tw, th = get_image_size(dst)
     assert tw <= width
     assert th <= height
@@ -162,7 +158,7 @@ def test_resize_upscale(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 500, 1000
-    resize_image(src, width, height, to=dst, method="cover")
+    resize_image(src, width, height, dst=dst, method="cover")
     tw, th = get_image_size(dst)
     assert tw == width
     assert th == height
@@ -176,7 +172,7 @@ def test_resize_small_image_error(png_image, jpg_image, tmp_path, fmt):
 
     width, height = 500, 1000
     with pytest.raises(ImageSizeError):
-        resize_image(src, width, height, to=dst, method="cover", allow_upscaling=False)
+        resize_image(src, width, height, dst=dst, method="cover", allow_upscaling=False)
 
 
 @pytest.mark.parametrize(
